+++ src/app/services/actions/query-action-creators.spec.ts
@@ -2,7 +2,7 @@
 import configureMockStore from 'redux-mock-store';
 
 import { ADD_HISTORY_ITEM_SUCCESS, QUERY_GRAPH_RUNNING, QUERY_GRAPH_STATUS, QUERY_GRAPH_SUCCESS } from '../redux-constants';
-import { runQuery } from '../slices/graphResponse.slice';
+import { runQuery } from '../slices/graph-response.slice';
 import { mockThunkMiddleware } from './mockThunkMiddleware';
 import { AnyAction } from '@reduxjs/toolkit';
 import { IQuery } from '../../../types/query-runner';
