+++ src/app/services/slices/graph-response.slice.ts
